Add files via upload
Initial upload, more comments are needed in the JS script. I started being tidy but code ended up a bit dirty. Just learned how to deal with setTimeout using arrays in order to deal with past values since values are updated and when setTimeout fires, it'll use actual values, not the ones I needed. To understand what I mean, try to show the value of a variable "i" inside a for loop, using setTimeout with 1000ms of delay. You need to store "i" variable inside an array. shift() and push() methods were used.